+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.8.1 - Unreleased
 
 - Added `text_keep` option to store extracted text on `IsccMeta.text` property
+- Changed pdf text extraction to use pypdfium2 with reading order reconstruction
 
 ## 0.8.0 - 2025-04-02